Tox: - added some debug info logging - fix for reading big files
git-svn-id: http://svn.miranda-ng.org/main/trunk@12083 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tox/src/tox_transfer.h
@@ -21,7 +21,7 @@ struct FileTransferParam
 	int friendNumber;
 	int fileNumber;
 
-	FileTransferParam(int friendNumber, int fileNumber, const TCHAR* fileName, size_t fileSize)
+	FileTransferParam(int friendNumber, int fileNumber, const TCHAR* fileName, uint64_t fileSize)
 	{
 		status = NONE;
 		hFile = NULL;
